feat: implement handler

+++ internal/author/repository.go
@@ -1,9 +1,9 @@
 package author
 
 type Repository interface {
-	GetByID(id int) (*Author, error)
+	GetByID(id int64) (*Author, error)
 	GetAll() ([]*Author, error)
 	Create(author *Author) error
 	Update(author *Author) error
-	Delete(id int) error
+	Delete(id int64) error
 }
