+++ internal/author/useCase.go
@@ -16,7 +16,7 @@ func (u *UseCase) Create(name string) (*Author, error) {
 	return author, err
 }
 
-func (u *UseCase) GetByID(id int) (*Author, error) {
+func (u *UseCase) GetByID(id int64) (*Author, error) {
 	return u.repo.GetByID(id)
 }
 
@@ -24,7 +24,7 @@ func (u *UseCase) GetAll() ([]*Author, error) {
 	return u.repo.GetAll()
 }
 
-func (u *UseCase) Update(id int, name string) (*Author, error) {
+func (u *UseCase) Update(id int64, name string) (*Author, error) {
 	author, err := u.repo.GetByID(id)
 	if err != nil {
 		return nil, err
@@ -40,11 +40,11 @@ func (u *UseCase) Update(id int, name string) (*Author, error) {
 	return author, nil
 }
 
-func (u *UseCase) Delete(id int) error {
+func (u *UseCase) Delete(id int64) error {
 	return u.repo.Delete(id)
 }
 
-func (u *UseCase) Deactivate(id int) (*Author, error) {
+func (u *UseCase) Deactivate(id int64) (*Author, error) {
 	author, err := u.repo.GetByID(id)
 	if err != nil {
 		return nil, err
@@ -59,7 +59,7 @@ func (u *UseCase) Deactivate(id int) (*Author, error) {
 	return author, nil
 }
 
-func (u *UseCase) Activate(id int) (*Author, error) {
+func (u *UseCase) Activate(id int64) (*Author, error) {
 	author, err := u.repo.GetByID(id)
 	if err != nil {
 		return nil, err
