+++ internal/post/entity.go
@@ -1,8 +1,8 @@
 package post
 
 type Post struct {
-	ID       int    `json:"id"`
+	ID       int64  `json:"id"`
 	Title    string `json:"title"`
 	Content  string `json:"content"`
-	AuthorID int    `json:"author_id"`
+	AuthorID int64  `json:"author_id"`
 }
