+++ internal/post/repository.go
@@ -1,9 +1,9 @@
 package post
 
 type Repository interface {
-	GetByID(id int) (*Post, error)
+	GetByID(id int64) (*Post, error)
 	GetAll() ([]*Post, error)
 	Create(post *Post) error
 	Update(post *Post) error
-	Delete(id int) error
+	Delete(id int64) error
 }
